Move any new types and interfaces to a `types.ts` to start being more consistent with the existing code structure conventions. Additionally remove a superfluous markdown link footer that came together strangely when accepted as a change in GitHub.
